langchain-main/requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
google-search-results
python-dotenv
